requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
ollama>=0.2.1
pydantic==2.6.4
httpx==0.27.0
sse-starlette==2.0.0
streamlit>=1.32.0
requests>=2.31.0
PyYAML>=6.0.1
